+++ internal/app/admin/consts/consts.go
@@ -6,4 +6,7 @@ const (
 
 	SysDictTypeStatusOk = "0"
 	SysDictTypeStatusNo = "1"
+
 )
+
+var PushChannelType = []string{"email", "webhook"}