+++ CHANGELOG.md
@@ -3,6 +3,6 @@
 ## [0.0.1] - 2025-01-15
 ### Init
 - Initial🎉
-- Douyin Recording(Cron, Monitor, Cookie, File)
-- Push Channel(Mail)
+- Douyin Recording(Cron, Monitor, Cookie, File, History)
+- Push Channel(Email)
 - System Management(User, Role, Menu, Config)